adds an end screen and better scoring and stuff

+++ Game2Boat/src/test/GameTest.java
@@ -33,7 +33,7 @@ public class GameTest {
 	public void lapTest(){
 		Game g1 = new Game();
 		g1.setLap(6);
-		assertEquals(6, g1.getLap());
+		assertEquals(6,(int) g1.getLap());
 	}
 
 }
